+++ full-stack/mern/mern-finance-tracker/client/src/pages/AccountsPage.jsx
@@ -1,12 +1,12 @@
 import { useState, useEffect, useCallback } from 'react'
 import { Plus, Wallet as WalletIcon } from 'lucide-react'
-import api from '../../lib/api'
-import { formatCurrency } from '../../lib/utils'
-import AccountCard from '../../components/features/AccountCard'
-import AccountForm from '../../components/features/AccountForm'
-import Modal from '../../components/common/Modal'
-import Loader from '../../components/common/Loader'
-import EmptyState from '../../components/common/EmptyState'
+import api from '../lib/api'
+import { formatCurrency } from '../lib/utils'
+import AccountCard from '../components/features/AccountCard'
+import AccountForm from '../components/features/AccountForm'
+import Modal from '../components/common/Modal'
+import Loader from '../components/common/Loader'
+import EmptyState from '../components/common/EmptyState'
 import toast from 'react-hot-toast'
 
 export default function AccountsPage() {
