+++ full-stack/mern/mern-finance-tracker/client/src/pages/BudgetsPage.jsx
@@ -1,12 +1,12 @@
 import { useState, useEffect, useCallback } from 'react'
 import { Plus, PiggyBank, Pencil, Trash2 } from 'lucide-react'
-import api from '../../lib/api'
-import { CATEGORY_LABELS } from '../../lib/utils'
-import Modal from '../../components/common/Modal'
-import BudgetForm from '../../components/features/BudgetForm'
-import BudgetProgressBar from '../../components/features/BudgetProgressBar'
-import Loader from '../../components/common/Loader'
-import EmptyState from '../../components/common/EmptyState'
+import api from '../lib/api'
+import { CATEGORY_LABELS } from '../lib/utils'
+import Modal from '../components/common/Modal'
+import BudgetForm from '../components/features/BudgetForm'
+import BudgetProgressBar from '../components/features/BudgetProgressBar'
+import Loader from '../components/common/Loader'
+import EmptyState from '../components/common/EmptyState'
 import toast from 'react-hot-toast'
 
 export default function BudgetsPage() {
