+++ full-stack/mern/mern-finance-tracker/client/src/pages/LoginPage.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 import { Link, useNavigate } from 'react-router-dom'
-import { useAuth } from '../../context/AuthContext'
+import { useAuth } from '../context/AuthContext'
 import { DollarSign, Eye, EyeOff } from 'lucide-react'
 import toast from 'react-hot-toast'
