+++ full-stack/mern/mern-finance-tracker/client/src/pages/ReportsPage.jsx
@@ -1,12 +1,12 @@
 import { useState, useEffect, useCallback } from 'react'
 import { BarChart3, Download, Calendar } from 'lucide-react'
-import api from '../../lib/api'
-import { formatCurrency, CATEGORY_LABELS, CATEGORY_COLORS } from '../../lib/utils'
-import TrendLineChart from '../../components/charts/TrendLineChart'
-import ExpenseBreakdownChart from '../../components/charts/ExpenseBreakdownChart'
-import IncomeVsExpenseChart from '../../components/charts/IncomeVsExpenseChart'
-import Loader from '../../components/common/Loader'
-import EmptyState from '../../components/common/EmptyState'
+import api from '../lib/api'
+import { formatCurrency, CATEGORY_LABELS, CATEGORY_COLORS } from '../lib/utils'
+import TrendLineChart from '../components/charts/TrendLineChart'
+import ExpenseBreakdownChart from '../components/charts/ExpenseBreakdownChart'
+import IncomeVsExpenseChart from '../components/charts/IncomeVsExpenseChart'
+import Loader from '../components/common/Loader'
+import EmptyState from '../components/common/EmptyState'
 import toast from 'react-hot-toast'
 
 export default function ReportsPage() {
